commented and updated readme
