+++ pax-compiler/src/building/apple.rs
@@ -1,10 +1,8 @@
 use colored::Colorize;
 use serde_json::Value;
 
-use crate::{
-    copy_dir_recursively, pre_exec_hook, wait_with_output, RunContext, RunTarget, BUILD_DIR_NAME,
-    DIR_IGNORE_LIST_MACOS, ERR_SPAWN, PAX_BADGE, PKG_DIR_NAME,
-};
+use crate::helpers::{BUILD_DIR_NAME, DIR_IGNORE_LIST_MACOS, ERR_SPAWN, PAX_BADGE, PKG_DIR_NAME};
+use crate::{copy_dir_recursively, pre_exec_hook, wait_with_output, RunContext, RunTarget};
 
 use color_eyre::eyre;
 use eyre::eyre;
