+++ pax-compiler/src/building/web.rs
@@ -1,9 +1,9 @@
 use crate::errors::source_map::SourceMap;
-use crate::helpers::wait_with_output;
-use crate::{
-    copy_dir_recursively, errors, pre_exec_hook, RunContext, RunTarget, ASSETS_DIR_NAME,
-    BUILD_DIR_NAME, DIR_IGNORE_LIST_WEB, ERR_SPAWN, PAX_BADGE, PKG_DIR_NAME, PUBLIC_DIR_NAME,
+use crate::helpers::{
+    wait_with_output, ASSETS_DIR_NAME, BUILD_DIR_NAME, DIR_IGNORE_LIST_WEB, ERR_SPAWN, PAX_BADGE,
+    PKG_DIR_NAME, PUBLIC_DIR_NAME,
 };
+use crate::{copy_dir_recursively, errors, pre_exec_hook, RunContext, RunTarget};
 
 use color_eyre::eyre;
 use colored::Colorize;
