+++ pax-compiler/src/manifest.rs
@@ -2,8 +2,8 @@ use std::collections::HashMap;
 use std::hash::Hasher;
 use std::{cmp::Ordering, hash::Hash};
 
+use crate::code_generation::templating::MappedString;
 use crate::parsing::escape_identifier;
-use crate::templating::MappedString;
 use serde_derive::{Deserialize, Serialize};
 #[allow(unused_imports)]
 use serde_json;
